+++ IntuneWinPackager.Models/Entities/DetectionProofMode.cs
@@ -1,8 +1,0 @@
-namespace IntuneWinPackager.Models.Entities;
-
-public enum DetectionProofMode
-{
-    PassiveRuleControl = 0,
-    ActiveInstallFlow = 1
-}
-